+++ app/src/main/java/com/neuralbit/letsnote/firebase/entities/TagFire.kt
@@ -1,4 +1,4 @@
-package com.neuralbit.letsnote.firebaseEntities
+package com.neuralbit.letsnote.firebase.entities
 
 data class TagFire(
     var tagName : String = "",
