+++ app/src/main/java/com/neuralbit/letsnote/firebase/repos/NoteFireRepo.kt
@@ -1,4 +1,4 @@
-package com.neuralbit.letsnote.firebaseRepos
+package com.neuralbit.letsnote.firebase.repos
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
@@ -8,8 +8,8 @@ import com.google.firebase.database.DatabaseError
 import com.google.firebase.database.ValueEventListener
 import com.google.firebase.database.ktx.database
 import com.google.firebase.ktx.Firebase
-import com.neuralbit.letsnote.firebaseEntities.NoteFire
-import com.neuralbit.letsnote.firebaseEntities.NoteFireIns
+import com.neuralbit.letsnote.firebase.entities.NoteFire
+import com.neuralbit.letsnote.firebase.entities.NoteFireIns
 import com.neuralbit.letsnote.utilities.NoteComparator
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
