+++ app/src/main/java/com/neuralbit/letsnote/firebase/repos/TagFireRepo.kt
@@ -1,4 +1,4 @@
-package com.neuralbit.letsnote.firebaseRepos
+package com.neuralbit.letsnote.firebase.repos
 
 import android.util.Log
 import androidx.lifecycle.LiveData
@@ -9,7 +9,7 @@ import com.google.firebase.database.DatabaseError
 import com.google.firebase.database.ValueEventListener
 import com.google.firebase.database.ktx.database
 import com.google.firebase.ktx.Firebase
-import com.neuralbit.letsnote.firebaseEntities.TagFire
+import com.neuralbit.letsnote.firebase.entities.TagFire
 
 class TagFireRepo {
