+++ app/src/main/java/com/neuralbit/letsnote/receivers/RescheduleAlarmsReceiver.kt
@@ -8,8 +8,8 @@ import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Context.ALARM_SERVICE
 import android.content.Intent
-import com.neuralbit.letsnote.firebaseEntities.NoteFire
-import com.neuralbit.letsnote.firebaseRepos.NoteFireRepo
+import com.neuralbit.letsnote.firebase.entities.NoteFire
+import com.neuralbit.letsnote.firebase.repos.NoteFireRepo
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
