+++ app/src/main/java/com/neuralbit/letsnote/ui/allNotes/AllNotesFragment.kt
@@ -34,7 +34,7 @@ import com.google.firebase.auth.FirebaseAuth
 import com.google.gson.Gson
 import com.neuralbit.letsnote.R
 import com.neuralbit.letsnote.databinding.FragmentAllNotesBinding
-import com.neuralbit.letsnote.firebaseEntities.NoteFire
+import com.neuralbit.letsnote.firebase.entities.NoteFire
 import com.neuralbit.letsnote.receivers.DeleteReceiver
 import com.neuralbit.letsnote.ui.adapters.NoteFireClick
 import com.neuralbit.letsnote.ui.adapters.NoteRVAdapter
