+++ app/src/main/java/com/neuralbit/letsnote/ui/archived/ArchivedViewModel.kt
@@ -6,7 +6,7 @@ import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.Transformations
-import com.neuralbit.letsnote.firebaseEntities.NoteFire
+import com.neuralbit.letsnote.firebase.entities.NoteFire
 import java.util.*
 
 class ArchivedViewModel (application : Application): AndroidViewModel(application) {
