+++ app/src/main/java/com/neuralbit/letsnote/ui/deletedNotes/DeletedNotesFragment.kt
@@ -25,7 +25,7 @@ import com.google.android.material.snackbar.Snackbar
 import com.google.gson.Gson
 import com.neuralbit.letsnote.R
 import com.neuralbit.letsnote.databinding.DeletedNotesFragmentBinding
-import com.neuralbit.letsnote.firebaseEntities.NoteFire
+import com.neuralbit.letsnote.firebase.entities.NoteFire
 import com.neuralbit.letsnote.receivers.DeleteReceiver
 import com.neuralbit.letsnote.ui.adapters.NoteFireClick
 import com.neuralbit.letsnote.ui.adapters.NoteRVAdapter
