+++ app/src/main/java/com/neuralbit/letsnote/ui/main/MainActivityViewModel.kt
@@ -5,12 +5,12 @@ import android.content.Context
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
-import com.neuralbit.letsnote.firebaseEntities.NoteFire
-import com.neuralbit.letsnote.firebaseEntities.TodoItem
-import com.neuralbit.letsnote.firebaseRepos.DeleteDataRepo
-import com.neuralbit.letsnote.firebaseRepos.LabelFireRepo
-import com.neuralbit.letsnote.firebaseRepos.NoteFireRepo
-import com.neuralbit.letsnote.firebaseRepos.TagFireRepo
+import com.neuralbit.letsnote.firebase.entities.NoteFire
+import com.neuralbit.letsnote.firebase.entities.TodoItem
+import com.neuralbit.letsnote.firebase.repos.DeleteDataRepo
+import com.neuralbit.letsnote.firebase.repos.LabelFireRepo
+import com.neuralbit.letsnote.firebase.repos.NoteFireRepo
+import com.neuralbit.letsnote.firebase.repos.TagFireRepo
 import com.neuralbit.letsnote.room.NoteDatabase
 import com.neuralbit.letsnote.room.relationships.NoteTagCrossRef
 import com.neuralbit.letsnote.room.repos.LabelRoomRepo
