+++ app/src/main/java/com/neuralbit/letsnote/ui/tag/TagViewModel.kt
@@ -3,7 +3,7 @@ package com.neuralbit.letsnote.ui.tag
 import android.app.Application
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.MutableLiveData
-import com.neuralbit.letsnote.firebaseEntities.TagFire
+import com.neuralbit.letsnote.firebase.entities.TagFire
 
 class TagViewModel (application: Application): AndroidViewModel(application) {
     var allTags  = ArrayList<Tag>()
