+++ app/src/main/java/com/neuralbit/letsnote/ui/todos/AllTodoFragment.kt
@@ -27,7 +27,7 @@ import com.google.android.material.floatingactionbutton.FloatingActionButton
 import com.google.gson.Gson
 import com.neuralbit.letsnote.R
 import com.neuralbit.letsnote.databinding.FragmentAllTodosBinding
-import com.neuralbit.letsnote.firebaseEntities.NoteFire
+import com.neuralbit.letsnote.firebase.entities.NoteFire
 import com.neuralbit.letsnote.receivers.DeleteReceiver
 import com.neuralbit.letsnote.ui.adapters.NoteFireClick
 import com.neuralbit.letsnote.ui.adapters.NoteRVAdapter
